+++ confeting/db/migrate/20160520142635_create_talks.rb
@@ -3,9 +3,9 @@ class CreateTalks < ActiveRecord::Migration
     create_table :talks do |t|
       t.string :title
       t.text :description
-      t.float :availability
-      t.float :conf_id
-      t.float :speaker_id
+      t.integer :availability
+      t.integer :conf_id
+      t.integer :speaker_id
 
       t.timestamps null: false
     end
